Tag UGens: Dfsm - m_count should be float
integers cannot be NaNs, required for demand-rate polling

+++ source/TagSystemUGens/TagSystemUgens.cpp
@@ -581,7 +581,7 @@ struct Dfsm : public Unit
 	int *m_nextstate_sizes;
 	int m_current_state;
 	int m_current_state_offset;
-	int m_count;
+	float m_count;
 	int m_end;
 };
